Extract MessageBus from SpoutCoordinator
